+++ src/commands/install.rs
@@ -1,3 +1,4 @@
+use std::clone;
 use crate::aliases::ModID;
 use crate::api::client::ApiClient;
 use crate::commands::sync::{get_sync_data};
@@ -9,33 +10,34 @@ use crate::version_management::{parse_latest_version};
 use std::path::PathBuf;
 use tracing::{debug, info};
 use crate::information_utils::{display_installation_results, notice};
+use crate::traits::ref_ext::PathRef;
 
 // Report if trying install a mod that already exists
 // Use -f to force an installation
 // add way to set the version you want to download
-pub async fn install_cmd(mod_dir: &PathBuf, mods_requested: Vec<ModID>, force: bool) -> Result<(), RustiqueError> {
-
+pub async fn install_cmd(mod_dir: impl PathRef, mods_requested: Vec<ModID>, force: bool) -> Result<(), RustiqueError> {
+    let mod_dir = mod_dir.as_ref();
     info!("install_cmd: {mods_requested:?}");
     // get sync data
     let sync_data = get_sync_data(mod_dir).await?;
 
     let installed_mods = sync_data.rustique_sync.clone();
-    // remove any mods from mods_requested if the exist in installed_mods
+    // remove any mods from mods_requested if they exist in installed_mods
 
-    let mods_requested_cleaned : Vec<ModID>  = mods_requested.iter().filter(|&id| !installed_mods.contains_key(id) && !force).cloned().collect();
-
-    if mods_requested.is_empty() {
-        notice("Looks like you have all the mods requested. If you would like to reinstall them, run this command again with --force", Some(comfy_table::Color::Yellow), vec![]);
-        return Err(SimpleError("No mods to install".to_string()))
-    }
+    // let mods_requested_cleaned : Vec<ModID>  = mods_requested.iter().filter(|&id| !installed_mods.contains_key(id) && !force).cloned().collect();
+    // 
+    // if mods_requested_cleaned.is_empty() {
+    //     notice("Looks like you have all the mods requested. If you would like to reinstall them, run this command again with --force", Some(comfy_table::Color::Yellow), vec![]);
+    //     return Err(SimpleError("No mods to install".to_string()))
+    // }
 
     let client = ApiClient::new();
 
     // get the download urls for all requested mods
-    let result = client.fetch_mods_parallel(mods_requested_cleaned.clone()).await?;
+    let result = client.fetch_mods_parallel(mods_requested.clone()).await?;
     
     if result.is_empty() {
-        return Err(SimpleError(format!("Invalid modid {mods_requested_cleaned:?}")));
+        return Err(SimpleError(format!("Invalid modid {mods_requested:?}")));
     }
 
     let mods_requested: Vec<Install> =
@@ -61,8 +63,8 @@ pub async fn install_cmd(mod_dir: &PathBuf, mods_requested: Vec<ModID>, force: b
 }
 
 
-pub async fn install_missing_deps(mod_dir: &PathBuf, mods_requested: Vec<ModID>) -> Result<(), RustiqueError> {
-
+pub async fn install_missing_deps<V: AsRef<[ModID]>>(mod_dir: impl PathRef, mods_requested: V) -> Result<(), RustiqueError> {
+    let (mod_dir , mods_requested) = (mod_dir.as_ref(), mods_requested.as_ref());
     // get all installed mod info
     // retrieve all dependencies
     // send missing ones to install_manager()