+++ src/rustique_errors.rs
@@ -31,6 +31,10 @@ pub enum RustiqueError {
     },
     ConfigFileError(String),
     MalformedModInfoJson(String),
+    TomlError{
+        context: String,
+        source: toml::de::Error
+    },
 
 }
 
@@ -49,6 +53,7 @@ impl fmt::Display for RustiqueError {
             RustiqueError::ModNotZipped(e) => write!(f, "Expected .zip, found folder. Did you forget to zip your mod? {}", e.to_string().yellow().bold()),
             RustiqueError::ConfigFileError(e) => write!(f, "Config File Error: {}", e.to_string().red().bold()),
             RustiqueError::MalformedModInfoJson(e) => write!(f, "Malformed modinfo.json discovered for {}: Please contact the mod author. Rustique cannot process this mod.", e.to_string().red().bold()),
+            RustiqueError::TomlError { context, source } => write!(f, "{}: {}", context, source.to_string().red().bold())
         }
     }
 }