+++ src/traits.rs
@@ -2,3 +2,4 @@ pub mod vec_ext;
 pub mod search_traits;
 pub mod option_ext;
 pub mod string_ext;
+pub mod ref_ext;