+++ src/traits/ref_ext.rs
@@ -1,0 +1,6 @@
+use std::path::Path;
+
+/// Trait for  AsRef\<Path\>
+pub trait PathRef: AsRef<Path> {}
+impl<T: AsRef<Path>> PathRef for T {}
+